Refactor Eth operations to their own container
