+++ src/containers/App/actions.js
@@ -2,6 +2,7 @@ import { SUPPORTED_NETWORKS } from 'config/constants';
 import {
   NOTIFY,
   CHANGE_NETWORK,
+  INIT_NETWORK_ACTIVITY,
 } from './constants';
 
 
@@ -18,5 +19,11 @@ export function changeNetwork(name) {
   return {
     type: CHANGE_NETWORK,
     network: SUPPORTED_NETWORKS[name],
+  };
+}
+
+export function initNetworkActivity() {
+  return {
+    type: INIT_NETWORK_ACTIVITY,
   };
 }
