+++ src/containers/App/constants.js
@@ -2,3 +2,5 @@ export const DEFAULT_LOCALE = 'en';
 
 export const NOTIFY = 'src/App/NOTIFY';
 export const CHANGE_NETWORK = 'src/App/CHANGE_NETWORK';
+
+export const INIT_NETWORK_ACTIVITY = 'src/App/INIT_NETWORK_ACTIVITY';
