+++ src/containers/EthOperationsHoc/constants.js
@@ -1,0 +1,10 @@
+/*
+ *
+ * EthOperationsHoc constants
+ *
+ */
+
+
+export const LOAD_BLOCK_HEIGHT = 'src/EthOperationsHoc/LOAD_BLOCK_HEIGHT';
+export const LOAD_BLOCK_HEIGHT_SUCCESS = 'src/EthOperationsHoc/LOAD_BLOCK_HEIGHT_SUCCESS';
+export const LOAD_BLOCK_HEIGHT_ERROR = 'src/EthOperationsHoc/LOAD_BLOCK_HEIGHT_ERROR';
