+++ src/containers/EthOperationsHoc/index.js
@@ -5,23 +5,17 @@
  */
 
 import React from 'react';
-import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
 
 import injectSaga from 'utils/injectSaga';
 import injectReducer from 'utils/injectReducer';
-import { initNetworkActivity } from 'containers/App/actions';
 import reducer from './reducer';
 import saga from './saga';
 
-function withHubiiApi(WrappedComponent) {
-  class HubiiApiHoc extends React.Component { // eslint-disable-line react/prefer-stateless-function
-    constructor(props) {
-      super(props);
-      this.props.initNetworkActivity();
-    }
+function withEthOperations(WrappedComponent) {
+  class EthOperationsHoc extends React.Component { // eslint-disable-line react/prefer-stateless-function
     render() {
       return (
         <div>
@@ -30,30 +24,24 @@ function withHubiiApi(WrappedComponent) {
       );
     }
   }
-
-  HubiiApiHoc.propTypes = {
-    initNetworkActivity: PropTypes.func.isRequired,
-  };
-
   const mapStateToProps = createStructuredSelector({});
 
   function mapDispatchToProps(dispatch) {
     return {
       dispatch,
-      initNetworkActivity: () => dispatch(initNetworkActivity()),
     };
   }
 
   const withConnect = connect(mapStateToProps, mapDispatchToProps);
 
-  const withReducer = injectReducer({ key: 'hubiiApiHoc', reducer });
-  const withSaga = injectSaga({ key: 'hubiiApiHoc', saga });
+  const withReducer = injectReducer({ key: 'ethOperationsHoc', reducer });
+  const withSaga = injectSaga({ key: 'ethOperationsHoc', saga });
 
   return compose(
     withReducer,
     withSaga,
     withConnect,
-  )(HubiiApiHoc);
+  )(EthOperationsHoc);
 }
 
-export default withHubiiApi;
+export default withEthOperations;
