+++ src/containers/EthOperationsHoc/selectors.js
@@ -1,0 +1,19 @@
+import { createSelector } from 'reselect';
+
+
+/**
+ * Direct selector to the ethOperationsHoc state domain
+ */
+const selectEthOperationsHocDomain = (state) => state.get('ethOperationsHoc');
+
+/**
+ * Other selectors
+ */
+const makeSelectBlockHeight = () => createSelector(
+  selectEthOperationsHocDomain,
+  (ethOperationsHocDomain) => ethOperationsHocDomain.get('blockHeight')
+);
+
+export {
+  makeSelectBlockHeight,
+};
