+++ src/containers/WalletHOC/constants.js
@@ -30,7 +30,3 @@ export const TRANSFER_ETHER = 'src/WalletHOC/TRANSFER_ETHER';
 export const TRANSFER_ERC20 = 'src/WalletHOC/TRANSFER_ERC20';
 export const TRANSFER_SUCCESS = 'src/WalletHOC/TRANSFER_SUCCESS';
 export const TRANSFER_ERROR = 'src/WalletHOC/TRANSFER_ERROR';
-
-export const LOAD_BLOCK_HEIGHT = 'src/WalletHOC/LOAD_BLOCK_HEIGHT';
-export const LOAD_BLOCK_HEIGHT_SUCCESS = 'src/WalletHOC/LOAD_BLOCK_HEIGHT_SUCCESS';
-export const LOAD_BLOCK_HEIGHT_ERROR = 'src/WalletHOC/LOAD_BLOCK_HEIGHT_ERROR';
