+++ src/mocks/store.js
@@ -10,9 +10,13 @@ import {
 import {
   appMock,
 } from 'containers/App/tests/mocks/selectors';
+import {
+  ethOperationsHocMock,
+} from 'containers/EthOperationsHoc/tests/mocks/selectors';
 
 export const storeMock = fromJS({
   app: appMock,
   hubiiApiHoc: hubiiApiHocMock,
   walletHoc: walletHocMock,
+  ethOperationsHoc: ethOperationsHocMock,
 });
